Debug progress
Not working yet, but on its way.
Extra \s in regex, silly me.
Forgot to typecast arguments to evalError(...) to str.
Changed "Line" to "Item" in evaluation errors, since that is really what
the index is.
Missed an evaluation case: when there are missing exits like ENDIF,
ENDWHILE, and DOWHILE.